requirements.txt
jieba==0.42.1
python-dotenv==1.0.1
python-telegram-bot==21.10
SQLAlchemy==2.0.27
